minor refactoring and code commenting
