+++ .gitignore
@@ -20,3 +20,4 @@ docs/_build
 population-500.xml
 /test/input/beam/sf-bay/model-inputs/otp/graphs
 /src/main/scala/beam/agentsim/config/BeamConfig.scala
+src/main/scala/beam/playground/zeeshan/
